Move vulkan.passes to vulkan._passes (#5919)
Summary:
Changing vulkan.passes to vulkan._passes to indicate that these passes are not covered under the API stability guarantee.

Pull Request resolved: https://github.com/pytorch/executorch/pull/5919

Reviewed By: helunwencser

Differential Revision: D63926849

fbshipit-source-id: bf135c46c6718bc37afa640cf51d004891516575
(cherry picked from commit e1832efa5544438cbcd981d43c6a10f254ecdafb)
